+++ src/Components/CreateImageWizard/steps/Oscap/removeBetaFromRelease.ts
@@ -1,0 +1,20 @@
+import {
+  RHEL_9_BETA,
+  RHEL_9,
+  RHEL_10_BETA,
+  RHEL_10,
+} from '../../../../constants';
+import { Distributions } from '../../../../store/imageBuilderApi';
+
+// The beta releases won't have any oscap profiles associated with them,
+// so just use the ones from the major release.
+export const removeBetaFromRelease = (dist: Distributions): Distributions => {
+  switch (dist) {
+    case RHEL_10_BETA:
+      return RHEL_10 as Distributions;
+    case RHEL_9_BETA:
+      return RHEL_9;
+    default:
+      return dist;
+  }
+};
